+++ src/ragger/navigator/navigator.py
@@ -17,7 +17,7 @@ from abc import ABC
 from pathlib import Path
 from tempfile import NamedTemporaryFile
 from time import time
-from typing import Callable, Dict, List, Optional, Union
+from typing import Callable, Dict, Optional, Sequence, Union
 
 from ragger.backend import BackendInterface, SpeculosBackend
 from ragger.firmware import Firmware
@@ -56,14 +56,16 @@ class Navigator(ABC):
         self._callbacks = callbacks
         self._golden_run = golden_run
 
-    def _get_snaps_dir_path(self, path: Path, test_case_name: Path, is_golden: bool) -> Path:
+    def _get_snaps_dir_path(self, path: Path, test_case_name: Union[Path, str],
+                            is_golden: bool) -> Path:
         if is_golden:
             subdir = "snapshots"
         else:
             subdir = "snapshots-tmp"
         return path / subdir / self._firmware.name / test_case_name
 
-    def _check_snaps_dir_path(self, path: Path, test_case_name: Path, is_golden: bool) -> Path:
+    def _check_snaps_dir_path(self, path: Path, test_case_name: Union[Path, str],
+                              is_golden: bool) -> Path:
         dir_path = self._get_snaps_dir_path(path, test_case_name, is_golden)
         if not dir_path.is_dir():
             if self._golden_run:
@@ -72,7 +74,10 @@ class Navigator(ABC):
                 raise ValueError(f"Golden snapshots directory ({dir_path}) does not exist.")
         return dir_path
 
-    def _init_snaps_temp_dir(self, path: Path, test_case_name: Path, start_idx: int = 0) -> Path:
+    def _init_snaps_temp_dir(self,
+                             path: Path,
+                             test_case_name: Union[Path, str],
+                             start_idx: int = 0) -> Path:
         snaps_tmp_path = self._get_snaps_dir_path(path, test_case_name, False)
         if snaps_tmp_path.exists():
             for file in snaps_tmp_path.iterdir():
@@ -141,7 +146,7 @@ class Navigator(ABC):
                          timeout: float = 10.0,
                          wait_for_screen_change: bool = True,
                          path: Optional[Path] = None,
-                         test_case_name: Optional[Path] = None,
+                         test_case_name: Optional[Union[Path, str]] = None,
                          snap_idx: int = 0) -> None:
         if isinstance(instruction, BaseNavInsID):
             instruction = NavIns(instruction)
@@ -208,8 +213,8 @@ class Navigator(ABC):
 
     def navigate_and_compare(self,
                              path: Optional[Path],
-                             test_case_name: Optional[Path],
-                             instructions: List[Union[InstructionType]],
+                             test_case_name: Optional[Union[Path, str]],
+                             instructions: Sequence[InstructionType],
                              timeout: float = 10.0,
                              screen_change_before_first_instruction: bool = True,
                              screen_change_after_last_instruction: bool = True,
@@ -221,10 +226,10 @@ class Navigator(ABC):
         :param path: Absolute path to the snapshots directory.
         :type path: Optional[Path]
         :param test_case_name: Relative path to the test case snapshots directory (from path).
-        :type test_case_name: Optional[Path]
+        :type test_case_name: Optional[Union[Path, str]]
         :param instructions: Set of navigation instructions. Navigation instruction IDs are also
                              accepted.
-        :type instructions: List[Union[NavIns, BaseNavInsID]]
+        :type instructions: Sequence[Union[NavIns, BaseNavInsID]]
         :param timeout: Timeout for each navigation step.
         :type timeout: int
         :param screen_change_before_first_instruction: Wait for a screen change before first
@@ -284,7 +289,7 @@ class Navigator(ABC):
         self._backend.resume_ticker()
 
     def navigate(self,
-                 instructions: List[InstructionType],
+                 instructions: Sequence[InstructionType],
                  timeout: float = 10.0,
                  screen_change_before_first_instruction: bool = True,
                  screen_change_after_last_instruction: bool = True) -> None:
@@ -294,7 +299,7 @@ class Navigator(ABC):
         :param instructions: Set of navigation instructions. Navigation instruction IDs are also
                              accepted, and will be converted into navigation instruction (without
                              any argument)
-        :type instructions: List[Union[NavIns, BaseNavInsID]]
+        :type instructions: Sequence[Union[NavIns, BaseNavInsID]]
         :param timeout: Timeout for each navigation step.
         :type timeout: int
         :param screen_change_before_first_instruction: Wait for a screen change before first
@@ -323,7 +328,7 @@ class Navigator(ABC):
                             navigate_instruction: InstructionType,
                             validation_instruction: InstructionType,
                             path: Path,
-                            test_case_name: Path,
+                            test_case_name: Union[Path, str],
                             start_img_name: str,
                             last_img_name: str,
                             take_snaps: bool = True,
@@ -349,7 +354,7 @@ class Navigator(ABC):
         :param path: Absolute path to the snapshots directory.
         :type path: Path
         :param test_case_name: Relative path to the test case snapshots directory (from path).
-        :type test_case_name: Path
+        :type test_case_name: Union[Path, str]
         :param start_img_name: Index of the first snapshot of the navigation flow.
         :type start_img_name: int
         :param last_img_name: Index of the snapshot to look for when navigating.
@@ -433,10 +438,10 @@ class Navigator(ABC):
 
     def navigate_until_text_and_compare(self,
                                         navigate_instruction: InstructionType,
-                                        validation_instructions: List[InstructionType],
+                                        validation_instructions: Sequence[InstructionType],
                                         text: str,
                                         path: Optional[Path] = None,
-                                        test_case_name: Optional[Path] = None,
+                                        test_case_name: Optional[Union[Path, str]] = None,
                                         timeout: int = 300,
                                         screen_change_before_first_instruction: bool = True,
                                         screen_change_after_last_instruction: bool = True) -> None:
@@ -452,15 +457,15 @@ class Navigator(ABC):
         action is performed on the device.
 
         :param path: Absolute path to the snapshots directory.
-        :type path: Path
+        :type path: Optional[Path]
         :param test_case_name: Relative path to the test case snapshots directory (from path).
-        :type test_case_name: Path
+        :type test_case_name: Optional[Union[Path, str]]
         :param navigate_instruction: Navigation instruction to be performed until the text is found.
                                      Navigation instruction ID is also accepted.
         :type navigate_instruction: Union[NavIns, BaseNavInsID]
         :param validation_instructions: Navigation instructions to be performed once the text is found.
                                         Navigation instruction IDs are also accepted.
-        :type validation_instructions: List[Union[NavIns, BaseNavInsID]]
+        :type validation_instructions: Sequence[Union[NavIns, BaseNavInsID]]
         :param text: Text string to look for.
         :type text: str
         :param timeout: Timeout for the whole navigation loop.
@@ -534,7 +539,7 @@ class Navigator(ABC):
 
     def navigate_until_text(self,
                             navigate_instruction: InstructionType,
-                            validation_instructions: List[InstructionType],
+                            validation_instructions: Sequence[InstructionType],
                             text: str,
                             timeout: int = 300,
                             screen_change_before_first_instruction: bool = True,
@@ -554,7 +559,7 @@ class Navigator(ABC):
         :type navigate_instruction: Union[NavIns, BaseNavInsID]
         :param validation_instructions: Navigation instructions to be performed once the text is found.
                                         Navigation instruction IDs are also accepted.
-        :type validation_instructions: List[Union[NavIns, BaseNavInsID]]
+        :type validation_instructions: Sequence[Union[NavIns, BaseNavInsID]]
         :param text: Text string to look for.
         :type text: str
         :param timeout: Timeout for the whole navigation loop.