+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.19.0] - 2024-05-14
+
+### Added
+- firmware/navigation: Adding swipe capabilities for Flex
+
+### Changed
+- navigation: scenarios now use `Firmware` enum rather than strings
+
 ## [1.18.2] - 2024-05-06
 
 ### Fixed